MDL-56317 quiz: adds converting auto-saves to standard saves
When enabled and  when an auto-save is processed if the last time a standard
save was completed on a question is longer than a Moodle administrator's
defined interval then it will perform a standard save instead.

This has been set as an interval instead of just always doing standard
saves to give administrators control over how much their database
grows while still allowing a high level of disaster recovery.