MDL-56317 quiz: adds question attempt response undo
This adds the ability for the user to pick a question attempt
state by time and apply it as a new question attempt answer state
while attempting an quiz.

This is disabled by default for questions and behaviours due to
the possibility of it creating unintended side-effects that are
not yet fully understood. This is currently available for
deferred feedback and manual graded behaviours and for the essay
type question.